broke up the app components into different files, easier to track each app component
