+++ app/src/main/java/com/ls/localsky/ui/components/TextComponents.kt
@@ -1,0 +1,131 @@
+package com.ls.localsky.ui.components
+
+
+import android.util.Log
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.heightIn
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.text.ClickableText
+import androidx.compose.material.Divider
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.SpanStyle
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.buildAnnotatedString
+import androidx.compose.ui.text.font.FontStyle
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.withStyle
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+
+@Composable
+
+fun TitleText(value: String) {
+    Text(
+        text= value,
+        modifier = Modifier
+            .fillMaxWidth()
+            .heightIn(min = 80.dp),
+        style = TextStyle(
+            fontSize = 34.sp,
+            fontWeight = FontWeight.Bold,
+            fontStyle = FontStyle.Normal,
+            textAlign = TextAlign.Center
+        )
+    )
+}
+
+@Composable
+fun DividerTextComponent() {
+    Row(
+        modifier = Modifier.fillMaxWidth(),
+        verticalAlignment = Alignment.CenterVertically
+    ){
+
+        Divider(modifier = Modifier
+            .fillMaxWidth()
+            .weight(1f),
+            color = Color.Gray,
+            thickness = 1.dp
+        )
+
+        Text(modifier = Modifier.padding(8.dp),
+            text= "or",
+            fontSize = 16.sp,
+            color = Color.Gray)
+
+        Divider(modifier = Modifier
+            .fillMaxWidth()
+            .weight(1f),
+            color = Color.Gray,
+            thickness = 1.dp
+        )
+    }
+}
+
+@Composable
+fun ClickableLoginText(onTextSelected : (String) -> Unit) {
+    val initText = "Already have an account? "
+    val loginText = "Login Here"
+
+    val annotatedString = buildAnnotatedString {
+        append(initText)
+        withStyle(style = SpanStyle(color = Color.Blue)) {
+            pushStringAnnotation(tag = loginText, annotation = loginText)
+            append(loginText)
+        }
+    }
+
+    ClickableText(
+        style = TextStyle(
+                textAlign = TextAlign.Center
+        ),
+        text = annotatedString, onClick = { offset ->
+        annotatedString.getStringAnnotations(offset, offset)
+            .firstOrNull()?.also{ span ->
+                Log.d("ClickableLoginText", "{${span.item}}")
+
+                if (span.item == loginText) {
+                    onTextSelected(span.item)
+                }
+
+            }
+    })
+
+}
+
+@Composable
+fun ClickableRegisterText(onTextSelected : (String) -> Unit) {
+    val initText = "Don't have an account? "
+    val registerText = "Register Here"
+
+    val annotatedString = buildAnnotatedString {
+        append(initText)
+        withStyle(style = SpanStyle(color = Color.Blue)) {
+            pushStringAnnotation(tag = registerText, annotation = registerText)
+            append(registerText)
+        }
+    }
+
+    ClickableText(
+        style = TextStyle(
+            textAlign = TextAlign.Center
+        ),
+        text = annotatedString, onClick = { offset ->
+        annotatedString.getStringAnnotations(offset, offset)
+            .firstOrNull()?.also{ span ->
+                Log.d("ClickableRegisterText", "{${span.item}}")
+
+                if (span.item == registerText) {
+                    onTextSelected(span.item)
+                }
+
+            }
+    })
+
+}
